add simple benchmarks
I thought that canonicalization was expensive, and I was right: adding a
canonicalization step makes finding the mount point about 7 times more
expensive in terms of time. I'd try to measure the comparison for memory
usage, as I suspect the additional allocations amount for a decent chunk
of the increased overhead, but that's still nontrivial when developing
Rust on macOS.

On my machine (2016 15" MBP with an Apple SSD and a 2.8GHz Core i7),
finding the mountpoint for a path is around 1μs, which is acceptable to
me.